DROPPING ITEMS WORKS.
everything is functional and ready to start with a basic game. next things to work on are a planting method and a way to get seed and whatnot. need to make an area for planting and a way for the seeds to be obtained. might need to make them into scriptable objects.